Move CoursierResolveKey to its own file to avoid a cycle with `compile.py`.
[ci skip-rust]
[ci skip-build-wheels]
