+++ src/python/pants/backend/scala/compile/scalac.py
@@ -22,12 +22,12 @@ from pants.jvm.resolve.coursier_fetch import (
     Coordinate,
     Coordinates,
     CoursierResolvedLockfile,
-    CoursierResolveKey,
     FilterDependenciesRequest,
     MaterializedClasspath,
     MaterializedClasspathRequest,
 )
 from pants.jvm.resolve.coursier_setup import Coursier
+from pants.jvm.resolve.key import CoursierResolveKey
 from pants.jvm.target_types import JvmArtifactFieldSet
 from pants.util.logging import LogLevel
